requirements.txt
#'tensorflow[and-cuda]' for cuda acceleration

deepface==0.0.93
Django==5.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
gunicorn==23.0.0
opencv-python==4.11.0.86
pgvector==0.4.0
pillow==11.2.1
psycopg==3.2.6
psycopg-binary==3.2.6
sounddevice==0.5.1
tensorflow==2.19.0
tensorflow-macos==2.16.2
tf_keras==2.19.0
ultralytics==8.3.111

